Initial support for translation Parallel Fortran Do loops using the default loop scheduling Added Fortran support for the following SageInterface functions * SageInterface::isMain() * SageInterface::setLoopLowerBound() * SageInterface::setLoopUpperBound() * SageInterface::setLoopStride()
New SageInterface functions:
* SageInterface::doLoopNormalization()
* SageInterface::isCanonicalDoLoop()
* SageInterface::findLastDeclarationStatement()

XOMP:
* changed long to int to accommodate Fortran integer type
* wrapper functions for XOMP_init(), XOMP_terminate(), etc

OpenMP lowering
* Added transOmpDo()

+++ src/midend/programTransformation/ompLowering/libxomp.h
@@ -40,7 +40,7 @@ extern void XOMP_taskwait (void);
 // scheduler initialization, only meaningful used for OMNI
 
 //Default loop scheduling, worksharing without any schedule clause
-extern void XOMP_loop_default(int lower, int upper, int stride, long * n_lower, long * n_upper);
+extern void XOMP_loop_default(int lower, int upper, int stride, int* n_lower, int* n_upper);
 
 extern void XOMP_loop_static_init(int lower, int upper, int stride, int chunk_size);
 extern void XOMP_loop_dynamic_init(int lower, int upper, int stride, int chunk_size);
